Started to bring all Scenes together
