--- src/test/java/com/molotkov/LoginSceneTest.java
@@ -1,4 +1,0 @@
-package com.molotkov;
-
-public class LoginSceneTest {
-}
