+++ src/test/java/com/molotkov/MainSceneTest.java
@@ -1,0 +1,4 @@
+package com.molotkov;
+
+public class MainSceneTest {
+}
